[Core] Project migration to CMake (#11)

+++ src/morse/morse.cpp
@@ -225,11 +225,11 @@ int main(int argc, char * argv[])
 	char cw[23];
 	float symbol_rate = wpm/1.25;
 
-	for (int i = 0; i < strlen(msg); i++)
+	for (size_t i = 0; i < strlen(msg); i++)
 	{
 		dits = text_to_morse(msg[i]);
 		morse_to_cw(dits, cw);
-		printf("msg[%02d]: %c\tmorse[%s]\tcw[%s]\n", i, toupper(msg[i]), dits, cw);
+		printf("msg[%02zu]: %c\tmorse[%s]\tcw[%s]\n", i, toupper(msg[i]), dits, cw);
 		Send_CW_OOK(freq, symbol_rate, cw);
 	}
 }
